# Parallelize `cc` and `bindgen` steps in backend build script