As of trunk c5082eadc0e4052b5655b33b89efb7f8c1b6d0c2, a run of `cargo +nightly build --timings` gives a 26.8s build on my machine. The long pole for this build is the `artichoke-backend` build script, which takes over 14 seconds to run.

`artichoke-backend` takes an additional 6 seconds to compile on top of the build script, so improvements we can make to speed up the build script have linear improvements on build time.

Rust 1.63.0, which Artichoke uses as its Rust toolchain as of #2076, introduces scoped threads to `std`. Scoped threads allow for parallelism which is scoped to a stack frame. This allows us to treat parallel execution similarly to an imperative style.

The `artichoke-backend` build script has two expensive halves:

- Compile mruby and mrbsys C sources into a static library using `cc`.
- Generate Rust code from mruby C headers using `bindgen`.

This patch spawns two threads using `std::thread::scope` to run these tasks in parallel. This single change makes the entire build of Artichoke from a clean state 35% faster. On my machine, build times dropped from 26.8s to 17.2s. The long pole of the build script dropped from 14.36s to 6.26s.

Timing reports from the compiler are attached to the PR associated with this patch.

## Timing Reports

### Before

**Timing Report**: [cargo-timing-20220829T014733Z.html.zip](https://github.com/artichoke/artichoke/files/9441096/cargo-timing-20220829T014733Z.html.zip)

Targets: | artichoke 0.1.0-pre.0 (lib, bin "airb", bin "artichoke")
-- | --
Profile: | dev
Fresh units: | 0
Dirty units: | 94
Total units: | 94
Max concurrency: | 16 (jobs=16 ncpu=16)
Build start: | 2022-08-29T01:47:33Z
Total time: | 26.8s
rustc: | rustc 1.65.0-nightly (c07a8b4e0 2022-08-26)<br/>Host: x86_64-apple-darwin<br/>Target: x86_64-apple-darwin
Max (global) rustc threads concurrency: | 0

<img width="334" alt="Screen Shot 2022-08-28 at 7 15 40 PM" src="https://user-images.githubusercontent.com/860434/187109805-78b1e650-54a7-48f6-a5ba-4808ccecfe2b.png">

### After

**Timing Report**: [cargo-timing-20220829T015620Z.html.zip](https://github.com/artichoke/artichoke/files/9441097/cargo-timing-20220829T015620Z.html.zip)

Targets: | artichoke 0.1.0-pre.0 (lib, bin "airb", bin "artichoke")
-- | --
Profile: | dev
Fresh units: | 0
Dirty units: | 94
Total units: | 94
Max concurrency: | 16 (jobs=16 ncpu=16)
Build start: | 2022-08-29T01:56:20Z
Total time: | 17.2s
rustc: | rustc 1.65.0-nightly (c07a8b4e0 2022-08-26)<br/>Host: x86_64-apple-darwin<br/>Target: x86_64-apple-darwin
Max (global) rustc threads concurrency: | 0

<img width="296" alt="Screen Shot 2022-08-28 at 7 16 58 PM" src="https://user-images.githubusercontent.com/860434/187109915-86ead3fb-d7df-4b6c-8365-7d7e5ab2358f.png">