@lopopolo 2022-08-29
From a cached state like CI builds from, it looks like compile times are roughly cut in half (from 60s to 30s). Overall CI build time is dominated by test runtime.